This is the initial commit of this fork of dagre. https://github.com/scottyantipa/dagre
The purpose of this fork is to solve some layout problems in the horizontal
compaction phase. Some layouts with the newer DFS algorithm (implemented in dagre)
would place nodes close to unrelated nodes in the same rank, while moving them
away from nodes that they are more closely related to (sharing a parent or child).

This commit does two things. 1) It undoes the changes to bk.js that were introduced
here: https://github.com/dagrejs/dagre/commit/bd1c10b4e6860056f139cbea96e0d55305100a03.
It does this by reverting bk.js to the state it was in the prior commit to the change
(eg: 7602bafe05340252e2c319b3f6bcc86852cad317). 2) We add a post-processing step after
calling `balance()` which detects if there are any collisions in positions for nodes in
the same rank. We added a naive algorithm to iteratively move any colliding nodes.