Ordering: Fix issue of overlapping node positions
This is a bug fix that addresses an issue where two nodes in the same
rank could end up having very close positions assigned causing them to
be overlapping. In the downstream application (knotend) this woud cause
a complete overlap because knotend calls Math.floor to descritize the
position. For example, dagre would emit two nodes with positions
4.5 and 4.2, which in a regular flowchart layout would cause an overlap
but not a complete overlap. Knotend maps those both to 4 causing them to
occupy the same cell in the grid, which is a complete bug.