Layout adjustment: Try to force tree structures to be compact
This is a knotend-specific fix. The original dagre does not try very
hard to keep siblings close together. For example if you have a tree,
the siblings of a certain node may get spread out from one another even
if they don't have any other connections at all (children or parents).
In knotend we want these nodes close together because it makes it easier
to select them with multi-select. It also just makes more sense
visually.

The algorithm works in two passes which are pretty identical. They
handle the two different cases of a) looking for tree strutures where
the siblings are on the left, eg they're sources of the main graph, and
b) looking for tree structes where the siblings are on the right, eg
they're sinks of the main graph. It tries to find these structures and
then sees if there's any space between the nodes that we can suck up. It
hen adjusts the positions of the nodes one by one to bring them closer
together without ever moving them over any other nodes that may be in
the way.