position.js: Add a phase that tries to bring siblings closer together
This is an attemp to fix a proble where dagre does not keep siblings
close together always in tree structures. For example, in a large tree,
especially towards the leaf nodes, some siblings will be placed very far
apart even if there is room between them to scrunch them together.

This commit adds a final pass that tries to walk these nodes closer
together. Note that this commit only implements this for tree structures
that point to the right -- eg, sub graphs which are themselves trees
who's leaf nodes are sinks in the original graph. We'll need to come
back and implement this for left tree structures, eg where the leaf
nodes are sources in the original graph.